Move remaining simplifiation into PlanarGraph class
