+++ prclz/reblock/i_topology_utils.py
@@ -73,86 +73,3 @@ def load_reblock_inputs(data_paths: Dict[str, Path],
     # Now load the parcels
     parcels_df = gpd.read_file(parcels_path)
     return parcels_df, buildings_df, blocks_df 
-
-# def update_edge_types(parcel_graph: PlanarGraph, 
-#                       block_polygon: Polygon, 
-#                       check: bool=False, 
-#                       lines_pgraph: PlanarGraph=None,
-#                       ) -> Tuple[int, int]:
-#     """
-#     When reblocking, existing roads receive 0 weight. Existing roads
-#     are captured in the block_polygon, so given a graph and a 
-#     block_geometry, this updates the edge types of the graph as
-#     either existing / new. Then it updates the weights.
-#     Graph is updated in place.
-
-#     Args:
-#         parcel_graph: graph repr of parcel boundaries
-#         block_geom: Polygon of block geometry
-#         check: If true, verify that each point in the block is in fact in the parcel
-#         lines_pgraph: NA - will be deprecated
-
-#     Returns:
-#         Updates graph in place. Returns summary of matching
-#         check between parcel and block
-#     """
-#     block_coords_list = list(block_polygon.exterior.coords)
-#     coords = set(block_coords_list)
-
-#     rv = (None, None)
-#     missing = None
-#     total = None
-    
-#     # Option to verify that each point in the block is in fact in the parcel
-#     if check:
-#         parcel_coords = set(v['name'] for v in parcel_graph.vs)
-#         total = 0
-#         is_in = 0
-#         for coord in coords:
-#             is_in = is_in+1 if coord in parcel_coords else is_in 
-#             total += 1
-#         missing = total-is_in
-#         if missing != 0:
-#             warning("{} of {} block coords are NOT in the parcel coords".format(missing, total)) 
-
-#     # Get list of coord_tuples from the polygon
-#     assert block_coords_list[0] == block_coords_list[-1], "Not a complete linear ring for polygon"
-
-#     # Loop over the block coords (as define an edge) and update the corresponding edge type in the graph accordingly
-#     # NOTE: every block coord will be within the parcel graph vertices
-#     for i, n0 in enumerate(block_coords_list):
-#         if i==0:
-#             continue
-#         else:
-#             n1 = block_coords_list[i-1]
-#             u_list = parcel_graph.vs.select(name_eq=n0)
-#             v_list = parcel_graph.vs.select(name_eq=n1)
-#             if len(u_list) > 0 and len(v_list) > 0:
-#                 u = u_list[0]
-#                 v = v_list[0]
-#                 path_idxs = parcel_graph.get_shortest_paths(u, v, weights='weight', output='epath')[0]
-
-#                 # the coords u and v from the block are
-#                 if lines_pgraph is None:
-#                     parcel_graph.es[path_idxs]['edge_type'] = 'highway'
-#                 else:
-#                     pass 
-#                     # ft_type = get_feature_type_from_lines(lines_pgraph, n0, n1 )
-#                     # parcel_graph.es[path_idxs]['edge_type'] = 'new'
-
-#                     # # Now view our new graph
-#                     # parcel_df = convert_to_gpd(parcel_graph)
-#                     # print("\n\n changing to {}".format(ft_type))
-#                     # plot_types(parcel_df)
-#                     # plt.show()
-#                     # parcel_graph.es[path_idxs]['edge_type'] = ft_type
-
-#     parcel_graph.es.select(edge_type_eq='highway')['weight'] = 0
-
-#     # WATERWAY_WEIGHT = NATURAL_WEIGHT = 1e4  # Not currently active
-#     # parcel_graph.es.select(edge_type_eq='waterway')['weight'] = WATERWAY_WEIGHT
-#     # parcel_graph.es.select(edge_type_eq='natural')['weight'] = NATURAL_WEIGHT
-
-#     rv = (missing, total)
-#     return rv 
-
